+++ Calculator/app/src/main/java/com/example/calculator_oop_kotlin_project/MathOperations.kt
@@ -4,52 +4,52 @@ import kotlin.math.pow
 
 class MathOperations() {
 
-    fun add(firstNumber: Double, secondNumber: Double): Double
+    fun add(firstNumber: Float, secondNumber: Float): Float
     {
         return firstNumber + secondNumber
     }
 
-    fun subtract(firstNumber: Double, secondNumber: Double): Double
+    fun subtract(firstNumber: Float, secondNumber: Float): Float
     {
         return firstNumber - secondNumber
     }
 
-    fun divide(firstNumber: Double, secondNumber: Double): Double
+    fun divide(firstNumber: Float, secondNumber: Float): Float
     {
         return firstNumber / secondNumber
     }
 
-    fun multiply(firstNumber: Double, secondNumber: Double): Double
+    fun multiply(firstNumber: Float, secondNumber: Float): Float
     {
         return firstNumber * secondNumber
     }
 
-    fun mod(firstNumber: Double, secondNumber: Double): Double
+    fun mod(firstNumber: Float, secondNumber: Float): Float
     {
         return firstNumber % secondNumber
     }
 
-    fun degree(firstNumber: Double, secondNumber: Double): Double
+    fun degree(firstNumber: Float, secondNumber: Float): Float
     {
         return firstNumber.pow(secondNumber)
     }
 
-    fun plusMinus(firstNumber: Double): Double
+    fun plusMinus(firstNumber: Float): Float
     {
         return firstNumber * -1
     }
 
-    fun reciprocal(firstNumber: Double): Double
+    fun reciprocal(firstNumber: Float): Float
     {
         return 1 / firstNumber
     }
 
-    fun sin(firstNumber: Double): Double
+    fun sin(firstNumber: Float): Float
     {
         return kotlin.math.sin(firstNumber)
     }
 
-    fun cos(firstNumber: Double): Double
+    fun cos(firstNumber: Float): Float
     {
         return kotlin.math.cos(firstNumber)
     }
